+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## next
 
+- new of() - utility function to create a one-shot stream of one value
+
 ## 0.8.1 - 2024-03-24
 
 - fix description for filter() transform